+++ src/CoreEx.Cosmos/Model/CosmosDbValueModelContainer.cs
@@ -41,7 +41,7 @@ namespace CoreEx.Cosmos.Model
         /// <param name="model">The model to infer <see cref="PartitionKey"/> from.</param>
         /// <param name="dbArgs">The <see cref="CosmosDbArgs"/>.</param>
         /// <returns>The <see cref="PartitionKey"/>.</returns>
-        /// <exception cref="AuthorizationException">Will be thrown where the infered <see cref="PartitionKey"/> is not equal to <see cref="CosmosDbContainerBase{TSelf}.DbArgs"/> (where not <c>null</c>).</exception>
+        /// <exception cref="AuthorizationException">Will be thrown where the infered <see cref="PartitionKey"/> is not equal to <see cref="CosmosDbContainer.DbArgs"/> (where not <c>null</c>).</exception>
         public PartitionKey GetPartitionKey(CosmosDbValue<TModel> model, CosmosDbArgs dbArgs)
         {
             var dbpk = DbArgs.PartitionKey;
@@ -117,7 +117,7 @@ namespace CoreEx.Cosmos.Model
         /// Gets the model for the specified <paramref name="key"/>.
         /// </summary>
         /// <param name="key">The <see cref="CompositeKey"/>.</param>
-        /// <param name="partitionKey">The <see cref="PartitionKey"/>. Defaults to <see cref="CosmosDbContainerBase{TSelf}.DbArgs"/>.</param>
+        /// <param name="partitionKey">The <see cref="PartitionKey"/>. Defaults to <see cref="CosmosDbContainer.DbArgs"/>.</param>
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
         /// <returns>The model value where found; otherwise, <c>null</c> (see <see cref="CosmosDbArgs.NullOnNotFound"/>).</returns>
         public async Task<CosmosDbValue<TModel>?> GetAsync(CompositeKey key, PartitionKey? partitionKey, CancellationToken cancellationToken = default) => await GetWithResultAsync(key, partitionKey, cancellationToken).ConfigureAwait(false);
@@ -126,7 +126,7 @@ namespace CoreEx.Cosmos.Model
         /// Gets the model for the specified <paramref name="key"/> with a <see cref="Result{T}"/>.
         /// </summary>
         /// <param name="key">The <see cref="CompositeKey"/>.</param>
-        /// <param name="partitionKey">The <see cref="PartitionKey"/>. Defaults to <see cref="CosmosDbContainerBase{TSelf}.DbArgs"/>.</param>
+        /// <param name="partitionKey">The <see cref="PartitionKey"/>. Defaults to <see cref="CosmosDbContainer.DbArgs"/>.</param>
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
         /// <returns>The model value where found; otherwise, <c>null</c> (see <see cref="CosmosDbArgs.NullOnNotFound"/>).</returns>
         public Task<Result<CosmosDbValue<TModel>?>> GetWithResultAsync(CompositeKey key, PartitionKey? partitionKey, CancellationToken cancellationToken = default) => GetWithResultAsync(new CosmosDbArgs(DbArgs, partitionKey), key, cancellationToken);
@@ -147,12 +147,12 @@ namespace CoreEx.Cosmos.Model
         /// <param name="key">The <see cref="CompositeKey"/>.</param>
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
         /// <returns>The model value where found; otherwise, <c>null</c> (see <see cref="CosmosDbArgs.NullOnNotFound"/>).</returns>
-        public Task<Result<CosmosDbValue<TModel>?>> GetWithResultAsync(CosmosDbArgs dbArgs, CompositeKey key, CancellationToken cancellationToken = default) => CosmosDb.Invoker.InvokeAsync(CosmosDb, GetCosmosId(key), dbArgs, async (_, key, args, ct) =>
+        public Task<Result<CosmosDbValue<TModel>?>> GetWithResultAsync(CosmosDbArgs dbArgs, CompositeKey key, CancellationToken cancellationToken = default) => CosmosDb.Invoker.InvokeAsync(CosmosDb, GetCosmosId(key), dbArgs, async (_, id, args, ct) =>
         {
             try
             {
                 var pk = args.PartitionKey ?? DbArgs.PartitionKey ?? PartitionKey.None;
-                var resp = await Container.ReadItemAsync<CosmosDbValue<TModel>>(key, pk, args.GetItemRequestOptions(), ct).ConfigureAwait(false);
+                var resp = await Container.ReadItemAsync<CosmosDbValue<TModel>>(id, pk, args.GetItemRequestOptions(), ct).ConfigureAwait(false);
                 if (resp.Resource == null || resp.Resource.Type != _typeName || args.FilterByTenantId && resp.Resource.Value is ITenantId tenantId && tenantId.TenantId != DbArgs.GetTenantId() || resp.Resource.Value is ILogicallyDeleted ld && ld.IsDeleted.HasValue && ld.IsDeleted.Value)
                     return args.NullOnNotFound ? Result<CosmosDbValue<TModel>?>.None : Result<CosmosDbValue<TModel>?>.NotFoundError();
 
@@ -258,9 +258,9 @@ namespace CoreEx.Cosmos.Model
 
             // Must read existing to update.
             ((ICosmosDbValue)m).PrepareBefore(dbArgs);
-            var key = m.Id;
+            var id = m.Id;
             var pk = GetPartitionKey(m, dbArgs);
-            var resp = await Container.ReadItemAsync<CosmosDbValue<TModel>>(key, pk, ro, ct).ConfigureAwait(false);
+            var resp = await Container.ReadItemAsync<CosmosDbValue<TModel>>(id, pk, ro, ct).ConfigureAwait(false);
             if (resp?.Resource == null || resp.Resource.Type != _typeName)
                 return Result<CosmosDbValue<TModel>>.NotFoundError();
 
@@ -282,7 +282,7 @@ namespace CoreEx.Cosmos.Model
                 })
                 .ThenAsAsync(async () =>
                 {
-                    resp = await Container.ReplaceItemAsync(resp.Resource, key, pk, ro, ct).ConfigureAwait(false);
+                    resp = await Container.ReplaceItemAsync(resp.Resource, id, pk, ro, ct).ConfigureAwait(false);
                     return GetResponseValue(resp)!;
                 });
         }, cancellationToken, nameof(UpdateWithResultAsync));
@@ -305,7 +305,7 @@ namespace CoreEx.Cosmos.Model
         /// Deletes the model for the specified <paramref name="key"/>.
         /// </summary>
         /// <param name="key">The <see cref="CompositeKey"/>.</param>
-        /// <param name="partitionKey">The <see cref="PartitionKey"/>. Defaults to <see cref="CosmosDbContainerBase{TSelf}.DbArgs"/>.</param>
+        /// <param name="partitionKey">The <see cref="PartitionKey"/>. Defaults to <see cref="CosmosDbContainer.DbArgs"/>.</param>
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
         public async Task DeleteAsync(CompositeKey key, PartitionKey? partitionKey, CancellationToken cancellationToken = default) => (await DeleteWithResultAsync(key, partitionKey, cancellationToken).ConfigureAwait(false)).ThrowOnError();
 
@@ -313,7 +313,7 @@ namespace CoreEx.Cosmos.Model
         /// Deletes the model for the specified <paramref name="key"/> with a <see cref="Result"/>.
         /// </summary>
         /// <param name="key">The <see cref="CompositeKey"/>.</param>
-        /// <param name="partitionKey">The <see cref="PartitionKey"/>. Defaults to <see cref="CosmosDbContainerBase{TSelf}.DbArgs"/>.</param>
+        /// <param name="partitionKey">The <see cref="PartitionKey"/>. Defaults to <see cref="CosmosDbContainer.DbArgs"/>.</param>
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
         public Task<Result> DeleteWithResultAsync(CompositeKey key, PartitionKey? partitionKey, CancellationToken cancellationToken = default) => DeleteWithResultAsync(new CosmosDbArgs(DbArgs, partitionKey), key, cancellationToken);
 
@@ -331,14 +331,14 @@ namespace CoreEx.Cosmos.Model
         /// <param name="dbArgs">The <see cref="CosmosDbArgs"/>.</param>
         /// <param name="key">The <see cref="CompositeKey"/>.</param>
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
-        public Task<Result> DeleteWithResultAsync(CosmosDbArgs dbArgs, CompositeKey key, CancellationToken cancellationToken = default) => CosmosDb.Invoker.InvokeAsync(CosmosDb, GetCosmosId(key), dbArgs, async (_, key, args, ct) =>
+        public Task<Result> DeleteWithResultAsync(CosmosDbArgs dbArgs, CompositeKey key, CancellationToken cancellationToken = default) => CosmosDb.Invoker.InvokeAsync(CosmosDb, GetCosmosId(key), dbArgs, async (_, id, args, ct) =>
         {
             try
             {
                 // Must read existing to delete and to make sure we are deleting for the correct Type; don't just trust the key.
                 var ro = args.GetItemRequestOptions();
                 var pk = args.PartitionKey ?? DbArgs.PartitionKey ?? PartitionKey.None;
-                var resp = await Container.ReadItemAsync<CosmosDbValue<TModel>>(key, pk, ro, ct).ConfigureAwait(false);
+                var resp = await Container.ReadItemAsync<CosmosDbValue<TModel>>(id, pk, ro, ct).ConfigureAwait(false);
                 if (resp?.Resource == null || resp.Resource.Type != _typeName)
                     return Result.Success;
 
@@ -357,7 +357,7 @@ namespace CoreEx.Cosmos.Model
                         .ThenAsync(async () =>
                         {
                             ro.SessionToken = resp.Headers?.Session;
-                            await Container.ReplaceItemAsync(resp.Resource, key, pk, ro, ct).ConfigureAwait(false);
+                            await Container.ReplaceItemAsync(resp.Resource, id, pk, ro, ct).ConfigureAwait(false);
                             return Result.Success;
                         });
                 }
@@ -367,7 +367,7 @@ namespace CoreEx.Cosmos.Model
                     .ThenAsync(async () =>
                     {
                         ro.SessionToken = resp.Headers?.Session;
-                        await Container.DeleteItemAsync<TModel>(key, pk, ro, ct).ConfigureAwait(false);
+                        await Container.DeleteItemAsync<TModel>(id, pk, ro, ct).ConfigureAwait(false);
                         return Result.Success;
                     });
             }
